+++ src/components/films/components/film-details/details-info/components/genre-list.js
