+++ src/components/films/components/film-details/film-details.js
@@ -1,6 +1,6 @@
-import {createDetailsInfo} from "./components/details-info";
-import {createControls} from "./components/controls";
-import {createCommentBlock} from "./components/comments";
+import {createDetailsInfo} from "./details-info/details-info";
+import {createControls} from "./constrols/controls";
+import {createCommentBlock} from "./comments/comments";
 
 
 /**
